Replace resource if timestamp is newer

+++ pyunity_editor/local.py
@@ -1,4 +1,5 @@
 from pathlib import Path
+from datetime import datetime
 import os
 import zipfile
 import shutil
@@ -15,13 +16,15 @@ directory = Path.home() / ".pyunity" / ".editor"
 
 def getPath(local):
     dest = directory / local
-    if dest.exists():
-        return str(dest)
     if egg:
         with zipfile.ZipFile(package) as zf:
             src = (Path(__package__) / local).as_posix()
             if src not in zf.namelist():
                 raise Exception(f"No resource at {package / src}")
+            if dest.exists():
+                ziptime = datetime(*zf.getinfo(src).date_time)
+                if ziptime.timestamp() != dest.stat().st_mtime:
+                    return dest
             out = zf.extract(src, directory)
             os.makedirs(dest.parent, exist_ok=True)
             shutil.move(out, dest.parent)
@@ -31,6 +34,8 @@ def getPath(local):
         src = package / local
         if not src.exists():
             raise Exception(f"No resource at {src}")
+        if dest.exists() and src.stat().st_mtime != dest.stat().st_mtime:
+            return dest
         dest.parent.mkdir(parents=True, exist_ok=True)
         shutil.copy(src, dest)
         return str(dest)
